decompress: change length arguments to long
In order to support decompression of files > 2GiB Linux has changed
the prototypes of decompression functions from

int uncompress(unsigned char *inbuf, int len,
	       int(*fill)(void*, unsigned int),
	       int(*flush)(void*, unsigned int),
	       unsigned char *output,
	       int *pos,
	       void(*error_fn)(char *x));

to

int uncompress(unsigned char *inbuf, long len,
	       long(*fill)(void*, unsigned long),
	       long(*flush)(void*, unsigned long),
	       unsigned char *output,
	       long *pos,
	       void(*error_fn)(char *x));

Do likewise in barebox for easier code sharing with Linux.

Link: https://lore.barebox.org/20240206094838.1987246-1-s.hauer@pengutronix.de

+++ include/linux/xz.h
@@ -264,10 +264,10 @@ XZ_EXTERN void xz_crc32_init(void);
 XZ_EXTERN uint32_t xz_crc32(const uint8_t *buf, size_t size, uint32_t crc);
 #endif
 
-STATIC int decompress_unxz(unsigned char *in, int in_size,
-		     int (*fill)(void *dest, unsigned int size),
-		     int (*flush)(void *src, unsigned int size),
-		     unsigned char *out, int *in_used,
+STATIC int decompress_unxz(unsigned char *in, long in_size,
+		     long (*fill)(void *dest, unsigned long size),
+		     long (*flush)(void *src, unsigned long size),
+		     unsigned char *out, long *in_used,
 		     void (*error)(char *x));
 
 #endif
